requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==4.0.1
biopython==1.86
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.128.7
fonttools==4.61.1
greenlet==3.3.1
h11==0.16.0
idna==3.11
kiwisolver==1.4.9
matplotlib==3.10.8
narwhals==2.16.0
numpy==2.4.2
packaging==26.0
passlib==1.7.4
pillow==12.1.1
plotly==6.5.2
pyasn1==0.6.2
pydantic==2.12.5
pydantic_core==2.41.5
PyMySQL==1.1.2
pyodbc==5.3.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.22
requests==2.32.5
rsa==4.9.1
scipy==1.17.0
six==1.17.0
SQLAlchemy==2.0.46
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
